feat(app): moves api endpoints into new routes structure
