+++ app/src/api/actions/login/useLoginUser.ts
@@ -1,10 +1,10 @@
 import { useMutation } from "@tanstack/react-query";
 import { loginUser } from ".";
-import { User } from "../../../types/user";
+import { Profile } from "../../../types/profile";
 
 export const useLoginUser = () => {
   return useMutation({
-    mutationFn: (user: Pick<User, "email" | "password">) => {
+    mutationFn: (user: Pick<Profile, "email" | "password">) => {
       return loginUser(user);
     },
   });
